refactor(api): split UserController into per-endpoint handlers
- Extract per-endpoint OpenAPI interfaces under features/users/api/openapi
- Add dedicated handler classes with unit tests for each endpoint
- Wire handlers in UserRoutes and DependencyContainer
- Move getParamId to common Utils and mock helpers to shared Helpers
- Remove UserController, UserApi, and UserControllerTest
