--- backend/api/src/main/java/com/anibalxyz/features/users/api/UserApi.java
@@ -1,218 +1,0 @@
-package com.anibalxyz.features.users.api;
-
-import com.anibalxyz.features.common.api.out.response.error.ErrorResponse;
-import com.anibalxyz.features.common.api.out.response.error.ErrorResponseExamples;
-import com.anibalxyz.features.users.api.in.UserCreateRequest;
-import com.anibalxyz.features.users.api.in.UserUpdateRequest;
-import com.anibalxyz.features.users.api.out.UserCreateResponse;
-import com.anibalxyz.features.users.api.out.UserDetailResponse;
-import com.anibalxyz.features.users.api.out.UsersErrorResponseExamples;
-import io.javalin.http.Context;
-import io.javalin.openapi.*;
-
-public interface UserApi {
-
-  @OpenApi(
-      summary = "Get all users",
-      operationId = "getAllUsers",
-      path = "/users",
-      methods = HttpMethod.GET,
-      tags = {"Users"},
-      security = @OpenApiSecurity(name = "bearerAuth"),
-      responses = {
-        @OpenApiResponse(
-            status = "200",
-            description = "A list of all users.",
-            content = @OpenApiContent(from = UserDetailResponse.Collection.class)),
-        @OpenApiResponse(
-            status = "401",
-            description = "Authentication information is missing or invalid.",
-            content =
-                @OpenApiContent(
-                    from = ErrorResponse.class,
-                    example = ErrorResponseExamples.UNAUTHORIZED)),
-        @OpenApiResponse(
-            status = "500",
-            description = "Internal server error.",
-            content =
-                @OpenApiContent(
-                    from = ErrorResponse.class,
-                    example = ErrorResponseExamples.INTERNAL_SERVER_ERROR))
-      })
-  void getAllUsers(Context ctx);
-
-  @OpenApi(
-      summary = "Get a user by ID",
-      operationId = "getUserById",
-      path = "/users/{id}",
-      methods = HttpMethod.GET,
-      tags = {"Users"},
-      security = @OpenApiSecurity(name = "bearerAuth"),
-      pathParams = {
-        @OpenApiParam(
-            name = "id",
-            type = Integer.class,
-            description = "The unique identifier of the user.",
-            required = true,
-            example = "1")
-      },
-      responses = {
-        @OpenApiResponse(
-            status = "200",
-            description = "Successfully retrieved the user.",
-            content = @OpenApiContent(from = UserDetailResponse.class)),
-        @OpenApiResponse(
-            status = "401",
-            description = "Authentication information is missing or invalid.",
-            content =
-                @OpenApiContent(
-                    from = ErrorResponse.class,
-                    example = ErrorResponseExamples.UNAUTHORIZED)),
-        @OpenApiResponse(
-            status = "400",
-            description = "Invalid ID format supplied.",
-            content =
-                @OpenApiContent(
-                    from = ErrorResponse.class,
-                    example = ErrorResponseExamples.INVALID_ID)),
-        @OpenApiResponse(
-            status = "404",
-            description = "User with the specified ID not found.",
-            content =
-                @OpenApiContent(
-                    from = ErrorResponse.class,
-                    example = ErrorResponseExamples.RESOURCE_NOT_FOUND))
-      })
-  void getUserById(Context ctx);
-
-  @OpenApi(
-      summary = "Create a new user",
-      operationId = "createUser",
-      path = "/users",
-      methods = HttpMethod.POST,
-      tags = {"Users"},
-      requestBody =
-          @OpenApiRequestBody(
-              description = "The user to create.",
-              required = true,
-              content = @OpenApiContent(from = UserCreateRequest.class)),
-      responses = {
-        @OpenApiResponse(
-            status = "201",
-            description = "User created successfully.",
-            content = @OpenApiContent(from = UserCreateResponse.class)),
-        @OpenApiResponse(
-            status = "400",
-            description =
-                "Invalid input provided, such as a duplicate email, missing fields, or invalid password format.",
-            content =
-                @OpenApiContent(
-                    from = ErrorResponse.class,
-                    example = UsersErrorResponseExamples.CREATE_USER_BAD_REQUEST)),
-        @OpenApiResponse(
-            status = "409",
-            description = "Conflict: email already in use.",
-            content =
-                @OpenApiContent(
-                    from = ErrorResponse.class,
-                    example = UsersErrorResponseExamples.EMAIL_ALREADY_IN_USE))
-      })
-  void createUser(Context ctx);
-
-  @OpenApi(
-      summary = "Update an existing user",
-      operationId = "updateUserById",
-      path = "/users/{id}",
-      methods = HttpMethod.PUT,
-      tags = {"Users"},
-      security = @OpenApiSecurity(name = "bearerAuth"),
-      pathParams = {
-        @OpenApiParam(
-            name = "id",
-            type = Integer.class,
-            description = "The unique identifier of the user to update.",
-            required = true,
-            example = "1")
-      },
-      requestBody =
-          @OpenApiRequestBody(
-              description = "The user data to update. At least one field must be provided.",
-              required = true,
-              content = @OpenApiContent(from = UserUpdateRequest.class)),
-      responses = {
-        @OpenApiResponse(
-            status = "200",
-            description = "User updated successfully.",
-            content = @OpenApiContent(from = UserDetailResponse.class)),
-        @OpenApiResponse(
-            status = "401",
-            description = "Authentication information is missing or invalid.",
-            content =
-                @OpenApiContent(
-                    from = ErrorResponse.class,
-                    example = ErrorResponseExamples.UNAUTHORIZED)),
-        @OpenApiResponse(
-            status = "400",
-            description = "Invalid input provided, such as empty payload, or duplicate email.",
-            content =
-                @OpenApiContent(
-                    from = ErrorResponse.class,
-                    example = UsersErrorResponseExamples.UPDATE_USER_BAD_REQUEST)),
-        @OpenApiResponse(
-            status = "404",
-            description = "User with the specified ID not found.",
-            content =
-                @OpenApiContent(
-                    from = ErrorResponse.class,
-                    example = ErrorResponseExamples.RESOURCE_NOT_FOUND)),
-        @OpenApiResponse(
-            status = "409",
-            description = "Conflict: email already in use.",
-            content =
-                @OpenApiContent(
-                    from = ErrorResponse.class,
-                    example = UsersErrorResponseExamples.EMAIL_ALREADY_IN_USE))
-      })
-  void updateUserById(Context ctx);
-
-  @OpenApi(
-      summary = "Delete a user by ID",
-      operationId = "deleteUserById",
-      path = "/users/{id}",
-      methods = HttpMethod.DELETE,
-      tags = {"Users"},
-      security = @OpenApiSecurity(name = "bearerAuth"),
-      pathParams = {
-        @OpenApiParam(
-            name = "id",
-            type = Integer.class,
-            description = "The unique identifier of the user to delete.",
-            required = true,
-            example = "1")
-      },
-      responses = {
-        @OpenApiResponse(status = "204", description = "User deleted successfully."),
-        @OpenApiResponse(
-            status = "401",
-            description = "Authentication information is missing or invalid.",
-            content =
-                @OpenApiContent(
-                    from = ErrorResponse.class,
-                    example = ErrorResponseExamples.UNAUTHORIZED)),
-        @OpenApiResponse(
-            status = "400",
-            description = "Invalid ID format supplied.",
-            content =
-                @OpenApiContent(
-                    from = ErrorResponse.class,
-                    example = ErrorResponseExamples.INVALID_ID)),
-        @OpenApiResponse(
-            status = "404",
-            description = "User with the specified ID not found.",
-            content =
-                @OpenApiContent(
-                    from = ErrorResponse.class,
-                    example = ErrorResponseExamples.RESOURCE_NOT_FOUND))
-      })
-  void deleteUserById(Context ctx);
-}
